refactor: Docs 수정 및 Alert 위치 수정
refactor: Docs 수정 및 Alert 위치 수정
