+++ src/main/java/com/team/comma/domain/playlist/playlist/repository/PlaylistRepositoryImpl.java
@@ -3,8 +3,7 @@ package com.team.comma.domain.playlist.playlist.repository;
 import com.querydsl.core.types.Projections;
 import com.querydsl.jpa.JPAExpressions;
 import com.querydsl.jpa.impl.JPAQueryFactory;
-import com.team.comma.domain.alert.dto.AlertResponse;
-import com.team.comma.domain.playlist.playlist.domain.Playlist;
+import com.team.comma.domain.playlist.alert.dto.AlertResponse;
 import com.team.comma.domain.playlist.playlist.dto.PlaylistResponse;
 import com.team.comma.domain.user.user.domain.User;
 import lombok.RequiredArgsConstructor;
